\name{morton-undocumented-beast}

\alias{beastIn}
\alias{sampleNodes}
\alias{summary.sampleNodes}

\title{undocumented morton functions - BEAST dating}

\description{\code{morton-undocumented} functions are currently undocumented}

\details{
\pkg{morton} is a work in progress. 
If you are interested in a function that is not yet documented, please contact Andrew Hipp (ahipp@mortonarb.org).
}

