\name{morton-undocumented-haversine}

\alias{haversine}
\alias{haversine.default}
\alias{haversine.matrix}
\alias{haversine.data.frame}


\title{undocumented morton functions - haversine}

\description{\code{morton-undocumented} functions are currently undocumented}

\details{
\pkg{morton} is a work in progress. 
If you are interested in a function that is not yet documented, please contact Andrew Hipp (ahipp@mortonarb.org).
## Arguments:
##  lat1 and long1: latitude and longitude for site one, in decimal format (e.g., N40deg 30', W 90deg 45' = 40.500, -90.750)
##  lat2 and long2: latitude and longitude for site two
##  r = radius of the earth in the units of interest. The default value is in Kilometers. This particular version of the formula assumes a spherical earth
## Andrew Hipp (ahipp@mortonarb.org), January 2008
}

