\name{morton-undocumented-mantel}

\alias{mantelMultiple}
\alias{Rsquared}
\alias{partialCorrelation}
\alias{permutedIndex}


\title{undocumented morton functions - mantel}

\description{\code{morton-undocumented} functions are currently undocumented}

\details{
\pkg{morton} is a work in progress. 
If you are interested in a function that is not yet documented, please contact Andrew Hipp (ahipp@mortonarb.org).
## Y must be a single distance matrix, the response
## X is a list of distance matrices, the predictors (can be a list of length == 1 for a standard Mantel test)
## The Pearson product moment is the hard-wired correlation coefficient as written (though of course any statistic could be used)
## Andrew Hipp, 22 May 2008 (ahipp@mortonarb.org)
}

