\name{morton-undocumented-other}

\alias{beta.phylo}
\alias{evenUp.phylo}
\alias{gls.r.squared}
\alias{tax2tree}
\alias{dist.nodes.2}

\title{undocumented morton functions - other}

\description{\code{morton-undocumented} functions are currently undocumented}

\details{
\pkg{morton} is a work in progress. 
If you are interested in a function that is not yet documented, please contact Andrew Hipp (ahipp@mortonarb.org).
}

