\name{morton-undocumented-rate.change}

\alias{rate.change}
\alias{rate.change.phylo}
\alias{rate.change.phylo4}
\alias{sharedBranches}
\alias{rootState}
\alias{sigmaSq}
\alias{logLik.phylogSigmaSq}

\title{undocumented morton functions - rate.change}

\description{\code{morton-undocumented} functions are currently undocumented}

\details{
\pkg{morton} is a work in progress. 
If you are interested in a function that is not yet documented, please contact Andrew Hipp (ahipp@mortonarb.org).

## new tree scaler -- equivalent to the non-censored O'Meara test
## the censored test has an extra degree of freedom, so the AIC 
## model approach swings away from detecting a change in rate
}

