\name{treeBuddy-package}
\alias{treeBuddy-package}
\alias{treeBuddy}
\docType{package}
\title{
\packageTitle{treeBuddy}
}
\description{
\packageDescription{treeBuddy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{treeBuddy}
\packageIndices{treeBuddy}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{treeBuddy}

Maintainer: \packageMaintainer{treeBuddy}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
